% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckdpc_ckd_risk_scores.R
\name{calculate_ckdpc_40egfr_risk}
\alias{calculate_ckdpc_40egfr_risk}
\title{Calculate CKDPC risk score for 3-year absolute risk of 40% decline in eGFR or kidney failure in people with diabetes with baseline eGFR>=60ml/min/1.73m2 (ckdpc_40egfr_risk) from Grams ME, Brunskill NJ, Ballew SH. Development and Validation of Prediction Models of Adverse Kidney Outcomes in the Population With and Without Diabetes Mellitus. Diabetes Care. doi:0.2337/dc22-0698 (https://diabetesjournals.org/care/article-abstract/45/9/2055/147251/Development-and-Validation-of-Prediction-Models-of?redirectedFrom=fulltext).}
\usage{
calculate_ckdpc_40egfr_risk(
  dataframe,
  age,
  sex,
  egfr,
  acr,
  sbp,
  bp_meds,
  hf,
  chd,
  af,
  current_smoker,
  ex_smoker,
  bmi,
  hba1c,
  oha,
  insulin
)
}
\arguments{
\item{dataframe}{dataframe containing variables for risk score}

\item{age}{current age in years}

\item{sex}{sex: "male" or "female"}

\item{egfr}{current eGFR in ml/min/1.73m2}

\item{acr}{albumin:creatinine ratio in mg/g}

\item{sbp}{systolic blood pressure in mmHg}

\item{bp_meds}{current use of antihypertensive medications (0: no, 1: yes)}

\item{hf}{history of heart failure (0: no, 1: yes)}

\item{chd}{history of coronary heart disease (MI or coronary revascularisation; 0: no, 1: yes)}

\item{af}{history of atrial fibrillation (0: no, 1: yes)}

\item{current_smoker}{current smoker (0: no, 1: yes)}

\item{ex_smoker}{ex_smoker (0: no, 1: yes)}

\item{bmi}{BMI in kg/m2}

\item{hba1c}{current HbA1c in mmol/mol}

\item{oha}{whether currently taking OHA (0: no, 1: yes)}

\item{insulin}{whether currently taking insulin (0: no, 1: yes)}
}
\description{
calculate 3-year absolute risk of 40% decline in eGFR or kidney failure in people with diabetes with baseline eGFR>=60ml/min/1.73m2
}
