% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ckdpcEgfr60RiskConstants}
\alias{ckdpcEgfr60RiskConstants}
\title{CKDPC eGFR<60 risk score constants}
\format{
An object of class \code{list} of length 21.
}
\usage{
ckdpcEgfr60RiskConstants
}
\description{
Constants required for to calculate 5-year CKDPC risk score for eGFR<60ml/min/1.73m2 (total and confirmed events) in people with diabetes. Source: Nelson RG, Grams ME, Ballew SH. Development of Risk Prediction Equations for Incident Chronic Kidney Disease. JAMA. doi:10.1001/jama.2019.17379 (https://jamanetwork.com/journals/jama/fullarticle/2755299).
}
\keyword{datasets}
