% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_cleaning.R
\name{clean_biomarker_units}
\alias{clean_biomarker_units}
\title{Clean biomarker units: only keep values with acceptable unit codes (includes missing unit code for all)}
\usage{
clean_biomarker_units(dataset, numunitid_col, biomrkr)
}
\arguments{
\item{dataset}{dataset containing biomarker observations}

\item{numunitid_col}{name of column containing numunitid codes}

\item{biomrkr}{name of biomarker to clean (acr/albumin_blood/alt/ast/bilirubin/bmi/creatinine_blood/dbp/fastingglucose/haemotocrit/haemoglobin/hba1c/hdl/height/ldl/pcr/sbp/totalcholesterol/triglyceride/weight)}
}
\description{
only keep measurements with 'acceptable' unit codes (numunitid) for specified biomarker
}
