% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qdiabeteshfConstants}
\alias{qdiabeteshfConstants}
\title{QDiabetes-Heart Failure constants}
\format{
An object of class \code{list} of length 2.
}
\usage{
qdiabeteshfConstants
}
\description{
Constants required for the QDiabetes-Heart Failure (2015) risk score. Source: https://qdiabetes.org/heart-failure/src.php
}
\keyword{datasets}
