% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qMissingPredictors}
\alias{qMissingPredictors}
\title{Q missing predictors}
\format{
An object of class \code{list} of length 2.
}
\usage{
qMissingPredictors
}
\description{
Constants required for to impute missing values for QRISK2 (2017) and QDiabetes-Heart Failure
}
\keyword{datasets}
