% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qrisk2Constants}
\alias{qrisk2Constants}
\title{QRISK2 constants}
\format{
An object of class \code{list} of length 2.
}
\usage{
qrisk2Constants
}
\description{
Constants required for the QRISK2 (2017) risk score. Source: https://www.qrisk.org/2017/QRISK2-2017-lgpl.tgz
}
\keyword{datasets}
